Extract debounce + sorry gink
