+++ data/ctfs.yml
@@ -2,7 +2,7 @@
   link: https://tjctf.org/
   startDate: 5/22/2020 11:00
   endDate: 5/26/2020 11:00
-- name: AngstromCTF 2020
+- name: ångstromCTF 2020
   link: https://2020.angstromctf.com/
   startDate: 3/14/2020 00:00
   endDate: 3/19/2020 00:00
