+++ src/utils/debounce.js
@@ -1,0 +1,14 @@
+const debounce = (func, wait) => {
+  let timeout
+
+  return (...args) => {
+    const later = () => {
+      clearTimeout(timeout)
+      func(...args)
+    }
+    clearTimeout(timeout)
+    timeout = setTimeout(later, wait)
+  }
+}
+
+export default debounce
